Move submitter tests to it's own file
